+++ modules/School_Setup/Configuration.php
@@ -291,7 +291,7 @@ else
 			// FJ add Registration to Configuration.
 			echo '<tr><td><br /><fieldset><legend>' . _( 'Registration' ) . '</legend><table>';
 
-			echo '<tr><td>' . CheckboxInput(
+			echo '<tr><td colspan="2">' . CheckboxInput(
 				Config( 'CREATE_USER_ACCOUNT' ),
 				'values[config][CREATE_USER_ACCOUNT]',
 				_( 'Create User Account' ) .
@@ -312,7 +312,7 @@ else
 					_( 'New students will be added as Inactive students' ) . '</i></div>';
 			}
 
-			echo '<tr><td>' . CheckboxInput(
+			echo '<tr><td colspan="2">' . CheckboxInput(
 				Config( 'CREATE_STUDENT_ACCOUNT' ),
 				'values[config][CREATE_STUDENT_ACCOUNT]',
 				_( 'Create Student Account' ) . $create_student_account_tooltip,
@@ -323,7 +323,8 @@ else
 			) . '</td></tr>';
 
 			// @since 5.9 Automatic Student Account Activation.
-			echo '<tr><td>' . CheckboxInput(
+			// HTML add arrow to indicate sub-option.
+			echo '<tr><td class="valign-top">&#10551; </td><td>' . CheckboxInput(
 				Config( 'CREATE_STUDENT_ACCOUNT_AUTOMATIC_ACTIVATION' ),
 				'values[config][CREATE_STUDENT_ACCOUNT_AUTOMATIC_ACTIVATION]',
 				_( 'Automatic Student Account Activation' ),
@@ -346,7 +347,8 @@ else
 			}
 
 			// @since 6.3 Create Student Account Default School.
-			echo '<tr><td>' . SelectInput(
+			// HTML add arrow to indicate sub-option.
+			echo '<tr><td class="valign-top">&#10551; </td><td>' . SelectInput(
 				Config( 'CREATE_STUDENT_ACCOUNT_DEFAULT_SCHOOL' ),
 				'values[config][CREATE_STUDENT_ACCOUNT_DEFAULT_SCHOOL]',
 				_( 'Default School' ),
@@ -365,7 +367,7 @@ else
 				$students_email_field_options[ $field['ID'] ] = ParseMLField( $field['TITLE'] );
 			}
 
-			echo '<tr><td>' . SelectInput(
+			echo '<tr><td colspan="2">' . SelectInput(
 				Config( 'STUDENTS_EMAIL_FIELD' ),
 				'values[config][STUDENTS_EMAIL_FIELD]',
 				_( 'Student email field' ),