+++ CHANGES.md
@@ -17,6 +17,7 @@ Changes in 11.2
 - Fix Attendance State Code: rename "Half Day" to "Half" in AttendanceCodes.php, AttendanceCodes.fnc.php & AttendanceSummary.php
 - CSS themes adjust tooltip height to fit inside PopTable in stylesheet.css & zresponsive.css
 - Fix SQL error when school has no MPs in Rollover.php
+- HTML add arrow to indicate sub-option in Configuration.php
 
 Changes in 11.1.2
 -----------------